separated stack from batch
